Update ddr3_controller.v
Microblaze used to throw "instruction over run" message when integrating with uberDDR3. The problem was that the DDR controller writes the entire RAM address space after calibration and reads it all back as part of test. This takes some time and while it's doing that Microblaze is ready to run and it immediately tries to access the RAM to start execution and stalls.
So, I skipped the part that writes the entire RAM address space after calibration since it can take quite sometime. 
After this small change, uberDDR3 works just like it did with MIG.